Move API metrics config validation into configuration parser
